More work on generic model

+++ layoutFX/dlTransfroms/SimpleTransformPane.java
@@ -215,7 +215,7 @@ public class SimpleTransformPane extends DLTransformPane {
 
 			}
 			else {
-				spinners.get(i).getValueFactory().setValue(simpleTransform.getParams()[i]);
+				spinners.get(i).getValueFactory().setValue(simpleTransform.getParams()[i].intValue());
 			}
 		}
